Day 12, Parts 1 and 2 solved
Notes:

Used an Enum for "Sides"
data Class for a "Plot", so we could eventually group plots into regions and determine which exact sides, if any, a given plot had fences on

- Part 1 was relatively easy, required a little thinking to get the region setting recursion working properly.

Was using a basic fence count (Int) for each Plot to make the price calculation easy.

- Part 2 required re-engineering of how a plots fences were being stored. Changed from a simple count, to storing which specific sides a plot had fences on, so we could calculate the sides.

The trick was figuring out that if we grouped all the plots in a region that had fences on a specific side, and then grouped by specific row/column, we could check if those grouped plots were adjacent, and therefore determine if they made one continuous side, or were multiple sides.

Getting better at recursion ;)