Day 15, Parts 1 and 2 solved
Notes:

Probably some room for refactoring in the bigBoxCanMove() function.

- Part 1 was fairly easy, although I tried to be too clever initially and use a data class for the individual squares, but it got very complicated with trying to find/move things, so I just reverted to a Map using the (x,y) co-ords as the Key and the SquareType as the value.

- Part 2 was easy to start with, expanding the board was relatively simple once I decided on the BOXLEFT/BOXRIGHT strategy, as I figured it would be helpful to know which part of a box we were trying to move

I then worked on getting left/right movement sorted (easy), as it was very similar to Part 1 movement.

Then tried the up/down movement (not so easy). It involved a lot more thinking and logic checks, but we got there in the end.

Creating the displayBoard() function to visually see what was happening to the boxes when I was trying to move them helped a lot at this point as I could see what was going wrong.