Day 16, Parts 1 and 2 solved
Notes:

Have finally seen the light and given up on Pair<Int, Int> for points ;)

Learnt about path/maze solving algorithms like Dijkstra and A*. Went with Dijkstra because it seemed to be the easiest to modify to return ALL the shortest paths.

Spent a bit of time trying to figure out how to build the initial graph. Specifically, the edge weights when turns were involved.

Managed to figure out corners/T-Junctions, but handling "crossroads" proved to be very difficult/"impossible". Instead, I ended up just implementing a "hack" in the Dijkstra Algorithm code to adjust the edge weight on the fly if we're turning at a crossroad.

Also modified the Dijkstra Algorithm so each point tracks its parents which form part of a "shortest path", so we can retrace the paths and identify all the unique Points for Part 2.