Day 17, Parts 1 and 2 solved
Notes:

Part 1 was fun and relatively straight forward. A chance to use some of the clever Kotlin tricks and techniques I have learned over the last few weeks.

Things like anonymous "first class" functions, lambda's, .map translations, "when" syntax

Code is probably not terribly idiomatic or performant, but it works ;)

Part 2 was a whole different kettle of fish. I initially tried brute-forcing the solution (it worked on the sample :P), but quickly realised that Ints weren't going to be big enough and I would need Longs, so the search space was going to be quite large.

I then tried to analyse exactly what the code was doing, and how we could "reverse engineer" it.

The first "aha!" moment was learning that dividing by 2^x is the same as a bitwise right shift by x bits. In other words, we are simply removing the last 3 bits from the number.

This meant, when working backwards, if we left shifted by 3 bits, we are effectively just adding 3 empty bits on the end.

If we then tried the various combinations of those missing 3 bits (ie. adding values 0 to 7 to the result of the left shift), we could effectively test the possible combinations of previous values of A, that once right shifted, we put us back to the value of A that we had prior to the left shift.

I was able to have this theory sort of confirmed from the Kotlin Slack. So I was then trying to figure out how the xor stuff could be "reversed", before realising, they were kind of a red herring.

Turns out, I could just iterate through A guesses, plug them into the program, and see if the resulting output, matched what we expected.

The trick was then realising I needed to try matching the program values, 1 step at a time, starting from the end. Increasing the number of program values to check by 1 each, each time we left shifted A.

Eventually got the algorithm locked in and correct values coming out ;)

Learned a lot about various programming shortcuts eg. divide by 2^x == shr, modulo x is the same as "and x-1", which can sometimes be more efficient etc.

Overall, a very challenging puzzle when you're not really up to speed on a lot of the "math tricks" with bitwise operations.