Day 19, Parts 1 and 2 solved
Notes:

Part 1 was relatively straightforward, some simple recursion through designs/partial designs checking if they could be constructed from available towel patterns. Returned the list of valid patterns, thinking it might help performance in Part 2.

For Part 2, I had a solution that worked for the small sample input, but would not complete in a timely manner due to the sheer volume of combinations available for the given designs.

Despite trying optimisations like only considering known valid designs from Part 1, and only trying to match towel patterns that are <= length of the (partial) design we are trying to match.

Looked at the Kotlin Slack, and was given a hint regarding "memoization" and went off to do some research. Basically, we need to cache the count values for the (partial) designs to avoid having to recalculate them repeatedly.

Another fun, learning day :)