Day 22, parts 1 and 2 solved
Notes:

After yesterday's puzzle, this was a breath of fresh air!

It's not a particularly elegant solution, as I believe there is built-in functionality for dealing with sequences and my (very) manual creation of such using .forEach (see fun getSequences()) is probably not very efficient (or required), but it works ;)

Did have a slight hiccup with Part 2, where I was taking the maximum price found for a sequence for a given buyer, rather than just the first price for the sequence.